stepcompress: Use 3rd-order Bézier velocity ramp.
This code uses a modified Newton-Raphson method to find the step times.
When there might be more than one root at some distance, it uses
gsl_poly_complex_solve() on the velocity function to find isolating
intervals for the roots of the position function. This way, the
(comparatively very expensive) GSL root finder is only used a maximum
of one time per stepcompress_push_bezier() call.

Only cartesian and corexy machines are supported for now.

You can use a 5th-order curve by changing BEZIER_ORDER to 5 in
stepcompress.c.

The GSL library is required in order to use this code.