extruder: Remove restriction regarding differing extrude ratios.
Previously, the planner forced a full stop when the extrusion width
changed. Theoretically, changing the extrusion width between moves requires
an instantaneous change in extruder position (obviously not possible). With
this commit, the position change occurs during the acceleration or
deceleration phase of the move. This is not really a very *good* solution
and could cause problems in certain cases, but may nonetheless be useful
until a better solution is implemented.